# internal/server/docs/swagger.yaml
components:
  schemas:
...
    api.APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        message:
          type: string
      type: object
...
    api.AddPropertyRequest:
      properties:
        description:
          type: string
        kind:
          type: string
        meta:
          additionalProperties:
            type: string
          description: |-
            Meta holds the consumer flags the server interprets — today only
            meta["index"] = "<scope>" | "none" for the search indexer
            (docs/13-index.md). Any other key is rejected; descriptive
            metadata lives under xFormat.
          type: object
        name:
          type: string
        scope:
          description: |-
            Scope is the property's write/sync class: "synced" (default,
            everyone in the space), "account" (this account's devices only),
            or "local" (this device only, never synced). "derived" is
            reserved for built-ins and rejected. Pinned by the first write,
            like kind — changing a property's scope means defining a new
            property. Property VALUE writes auto-route by the declared scope
            (POST /v1/spaces/:spaceId/properties/:objectId/set/:typeId).
          type: string
        xFormat:
          additionalProperties: {}
          description: |-
            XFormat is the property's descriptor: everything descriptive
            beyond the kind — the semantic slug, icon, ordering key, option
            set, relation targets, per-format config (docs/27-descriptors.md).
            An object; the server validates the keys it interprets (type,
            icon, pos, options, relation, config, links) against the vocabulary and
            the slug against kind, stores vendor-namespaced keys verbatim,
            and reserves validate / compute. Every path under it is mutable
            via PATCH.
          type: object
        xKey:
          description: |-
            XKey is the property's handle — an alias, not a storage key
            (values live under the content-addressed propId). Unique within
            the type (409 property.xkey_conflict); mutable via PATCH.
          type: string
      type: object
    api.AddPropertyResponse:
      properties:
        propId:
          type: string
      type: object
...
    api.CollectionInfo:
      properties:
        builtIn:
          type: boolean
        description:
          type: string
        hidden:
          type: boolean
        iconCid:
          type: string
        id:
          type: string
        meta:
          additionalProperties: {}
          type: object
        name:
          type: string
        xKey:
          type: string
      type: object
    api.CollectionPatchRequest:
      properties:
        description:
          type: string
        hidden:
          type: boolean
        iconCid:
          type: string
        meta:
          additionalProperties: {}
          type: object
        name:
          type: string
      type: object
    api.CollectionsCreateRequest:
      properties:
        description:
          type: string
        hidden:
          type: boolean
        iconCid:
          type: string
        meta:
          additionalProperties: {}
          type: object
        name:
          type: string
        xKey:
          type: string
      type: object
    api.CollectionsCreateResponse:
      properties:
        collectionId:
          type: string
      type: object
    api.CollectionsListResponse:
      properties:
        collections:
          items:
            $ref: '#/components/schemas/api.CollectionInfo'
          type: array
          uniqueItems: false
      type: object
...
    api.ErrorEnvelope:
      properties:
        error:
          $ref: '#/components/schemas/api.APIError'
      type: object
...
    api.PropertiesListResponse:
      properties:
        properties:
          items:
            $ref: '#/components/schemas/api.PropertyDef'
          type: array
          uniqueItems: false
      type: object
...
    api.PropertyDef:
      properties:
        description:
          type: string
        id:
          type: string
        items:
          $ref: '#/components/schemas/api.PropertyDef'
        kind:
          type: string
        meta:
          additionalProperties:
            type: string
          description: |-
            Meta is the consumer flag map (meta["index"] for the search
            indexer).
          type: object
        name:
          type: string
        properties:
          items:
            $ref: '#/components/schemas/api.PropertyDef'
          type: array
          uniqueItems: false
        required:
          items:
            type: string
          type: array
          uniqueItems: false
        scope:
          description: |-
            Scope is the property's write/sync class (synced / derived /
            account / local). Definitions written before scopes existed
            read back as "synced".
          type: string
        xFormat:
          additionalProperties: {}
          description: |-
            XFormat is the descriptor as stored — absent for a property that
            never declared one, which renders structurally from Kind.
          type: object
        xKey:
          type: string
      type: object
    api.PropertyPatchRequest:
      properties:
        set:
          type: object
        unset:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
info:
  description: Local HTTP/JSON API wrapping any-sync-sdk. Localhost-only, no auth
    in v1.
  title: Any API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /spaces/{spaceId}/collections:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Include hidden collections
        in: query
        name: includeHidden
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.CollectionsListResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: List collections
      tags:
      - collections
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.CollectionsCreateRequest'
                description: Collection params
                summary: body
        description: Collection params
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.CollectionsCreateResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Create a collection
      tags:
      - collections
  /spaces/{spaceId}/collections/{collectionId}:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.CollectionInfo'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Get a collection
      tags:
      - collections
    patch:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.CollectionPatchRequest'
                description: Fields to change
                summary: body
        description: Fields to change
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Patch a collection's display metadata
      tags:
      - collections
  /spaces/{spaceId}/collections/{collectionId}/properties:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.PropertiesListResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: List properties of a collection
      tags:
      - collections
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.AddPropertyRequest'
                description: Property draft
                summary: body
        description: Property draft
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.AddPropertyResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Add a property to a collection
      tags:
      - collections
  /spaces/{spaceId}/collections/{collectionId}/properties/{propId}:
    delete:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      - description: Property ID
        in: path
        name: propId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Remove a property from a collection
      tags:
      - collections
    patch:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      - description: Property ID
        in: path
        name: propId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.PropertyPatchRequest'
                description: Patch
                summary: body
        description: Patch
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Patch a property of a collection
      tags:
      - collections
...
servers:
- url: http://127.0.0.1:7001/v1